+++ src/main/java/lv/lottery/dto/StopRegistrationDTO.java
@@ -1,17 +1,17 @@
 package lv.lottery.dto;
 
 public class StopRegistrationDTO {
-    private Long id;
+    private Integer id;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
-    public StopRegistrationDTO(Long id) {
+    public StopRegistrationDTO(Integer id) {
         this.id = id;
     }
 }
